chore(friendshipper): move handleAutoLaunch into ServerModal
